chore(rpc): separate benchmarks from tests
