+++ internal/rpc/engine_test.go
@@ -1,10 +1,8 @@
 package rpc
 
 import (
-	"bytes"
 	"errors"
 	"sync"
-	"sync/atomic"
 	"testing"
 	"time"
 
@@ -461,67 +459,4 @@ func runTest(
 	assert.NoError(t, g.Wait())
 	e.Close()
 	assert.NoError(t, cfg.Logger.Sync())
-}
-
-// mockObject implements bin.Object for testing.
-type mockObject struct {
-	data []byte
-}
-
-func (m mockObject) Decode(b *bin.Buffer) error {
-	if !bytes.Equal(b.Buf, m.data) {
-		return errors.New("mismatch")
-	}
-	b.Skip(len(b.Buf))
-	return nil
-}
-
-func (m mockObject) Encode(b *bin.Buffer) error {
-	b.Put(m.data)
-	return nil
-}
-
-func BenchmarkEngine_Do(b *testing.B) {
-	ids := make(chan int64, 100)
-	defer close(ids)
-
-	e := New(func(ctx context.Context, msgID int64, seqNo int32, in bin.Encoder) error {
-		ids <- msgID
-		return nil
-	}, Options{})
-
-	var id int64
-
-	ctx := context.Background()
-	b.ReportAllocs()
-	b.RunParallel(func(pb *testing.PB) {
-		go func() {
-			buf := &bin.Buffer{}
-			// Fake handler.
-			obj := mockObject{data: make([]byte, 100)}
-
-			for id := range ids {
-				e.NotifyAcks([]int64{id})
-
-				buf.ResetTo(obj.data)
-				if err := e.NotifyResult(id, buf); err != nil {
-					b.Error(err)
-				}
-			}
-		}()
-
-		obj := mockObject{data: make([]byte, 100)}
-
-		for pb.Next() {
-			nextID := atomic.AddInt64(&id, 1)
-			if err := e.Do(ctx, Request{
-				ID:       nextID,
-				Sequence: int32(nextID),
-				Input:    obj,
-				Output:   obj,
-			}); err != nil {
-				b.Fatal(err)
-			}
-		}
-	})
 }
